rename dispatch to algorithm
