+++ crates/cubecl-linalg/src/matmul/kernels/cmma_matmul/algorithm/mod.rs
@@ -1,0 +1,6 @@
+mod base;
+
+pub mod cmma;
+pub mod plane_mma;
+
+pub use base::Algorithm;
